requirements.txt
# Data Handling & Analysis
pandas>=2.0.0
numpy>=1.26.4
pyarrow>=16.1.0

# Machine Learning
scikit-learn>=1.7.0
xgboost>=2.0.3
lightgbm>=4.0.0
imbalanced-learn>=0.10.0
joblib>=1.4.2

# Technical Analysis
pandas-ta==0.3.14b0

# Database & Environment
influxdb-client[ciso]>=1.42.0
python-dotenv>=1.0.1

# Visualization
matplotlib>=3.8.4
